Color.h: A Color object now takes up 2 bytes instead of 4, because we only need to cover the range [-3, 255] and uint8_t doesn't quite cut it because of the negative special color IDs, alas.

+++ include/Termin8or/screen/Color.h
@@ -717,7 +717,7 @@ namespace t8
     }
     
   private:
-    int idx = 0;
+    int16_t idx = 0; // Range: [-3, 255]
     
     void set_color16(Color16 col16)
     {
